requirements.txt
absl-py==1.0.0
asttokens==2.0.5
astunparse==1.6.3
attrdict==2.0.1
backcall==0.2.0
bdlb @ git+ssh://git@github.com/hermannsblum/bdl-benchmark.git@03773d621374ed10122866e64943df77ff8fbf50
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi==2022.6.15
charset-normalizer==2.0.12
clip @ git+ssh://git@github.com/openai/CLIP.git@b46f5ac7587d2e1862f8b7b1573179d80dcdd620
cycler==0.11.0
dataclasses==0.6
debugpy==1.5.1
decorator==5.1.1
dill==0.3.4
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
flatbuffers==2.0
fonttools==4.33.3
ftfy==6.1.1
future==0.18.2
gast==0.5.3
gdown==4.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.0
grpcio==1.46.0
h5py==3.6.0
idna==3.3
imageio==2.9.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.9.1
ipython==8.3.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.2.2
jupyter-core==4.10.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
kornia @ git+ssh://git@github.com/kornia/kornia@af948c5baddbbe1b4c4f1cde86817a4408fe7167
libclang==14.0.1
llvmlite==0.38.0
Markdown==3.3.7
matplotlib==3.5.2
matplotlib-inline==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
nest-asyncio==1.5.5
networkx==2.8
numpy==1.22.3
oauthlib==3.2.0
ood-metrics==0.2.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
pyzmq==22.3.0
regex==2022.6.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.16.2
scikit-learn==0.24.1
scipy==1.8.0
setuptools==61.2.0
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4
tensorflow==2.8.0
tensorflow-datasets==4.5.2
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-metadata==1.8.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
timm==0.6.7
tornado==6.1
tqdm==4.61.0
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
wheel==0.37.1
wrapt==1.14.1
zipp==3.8.0
